fix SWAYSOCK environment variable management
After a logout and new login, a new sway instance is created with a
different socket. Existing tmux sessions and zsh shells would still have
the previous variable value when reattached, making it impossible to use
swaymsg commands for example. Another issue was in reloading a shell
which would update systemd environment to the old variable.

Now the following is done:
- reattaching a tmux session updates the SWAYSOCK variable to new value
- reloading shells updates the variable from the tmux value
- systemd environment update is done only on login shells