remove tmux alert message display restriction
First we need the desktop notification displayed even if the attached
session and active window is the same, in case we are in a different
workspace.
Even if we stay on the same window, the pane where the alert is coming
from might be hidden from another one being zoomed in, so the message
displayed internally would still be useful.

Better always displayed both than adding complex logic here.
The disturbance it can create would be small and quite rare.
"Worse is better"