decrease tmux key repeat time to prevent wrong binding being triggered
All keybindings using the repeat flag are possible to be used once such
a keybinding has been pressed, while I thought this was limited only to
that one key being pressed.

The main problem happens in the following scenario:
- press select next window binding (prefix+n with repeat flag)
- press ctrl-p in the shell of that new window to browse history

If we press ctrl-p too quickly (while still during the
repeat time interval), the select previous session binding is triggered
instead (prefix+C-p with repeat flag).

The repeat flag seems to leave the prefix mode active for the repeat
time interval for any other binding with that flag to be used.

The behavior was like this in previous tmux versions but somehow such
issues started to become visible only in new version (?).

Decreasing the repeat time interval mitigates the issue:
- long enough to still allow new keys to be pressed
- short enough to prevent errors or waiting to start other interactions