make all sway app_* scripts able to reattach existing tmux sessions
The window is gone so it can't be focused on but the app might still be
running in a background tmux session if it was just detached instead of
killed.

We preferably want to reattach the existing session instead of launching
a new one. For weechat it was actually failing since it is not possible
to launch multiple instances of it.

And this way we can now delete the chat window while letting the app run
in the background and mainly relies on the notifications it triggers.