fix qutebrowser adblock systemd timer
Using both OnCalendar and onBootSec doesn't seem to do what was wanted
meaning a weekly run 1H after booting when triggered. Instead the run
happens 1H after every boot (and maybe weekly if the system would be left
running for so long)
Persistent usage was also problematic since it meant the run would
happen automatically on activation when needed, before the DISPLAY is
available.

There doesn't seem to be a simple way to do what was wanted so we now
just fetch the adblock list after every boot even if it is a bit overkill.
We move the timing to 4H so it doesn't happen for just short sessions.