ignore floating windows when creating a new sway workspace
if create workspace is called from a launcher, this one is still
detected as being present in the current workspace, which is then marked
as used even if no other window is there.

The use we have of floating windows is only for temporary ones so it's
alright not to count them as using a workspace.