turn script --help usage display into proper man pages
Documentation is now disconnected from the script code but this is not a
big issue as there was already a necessary separation between the code
and the completion file for instance.
Having the doc within the source file didn't guarantee a regular update
anyway, just made it more likely.

The man pages allows for better documentation.
They also are expected for "git help _command_" display too.

utilities doc have not been turned into man pages as those scripts are
expected to be only run locally within the repo