remove or replace bindings using Alt in zsh
Alt bindings are also triggered via Esc, so the action sometimes get
triggered by mistake when switching to command mode and typing the next
key too fast.
Typically Esc + dd to delete the line was often triggering the switch to
Desktop action.

root and desktop bindings are removed because not that much useful
reload binding is replaced with an alias (like others)
edit command line uses Ctrl instead of Alt