Build documentation cache via systemd timer
Daily packages update forced the rebuild of the cache, most of the time
needlessly (no new doc or executable, just file update). This was rather
annoying to have to wait to access the doc needed.

Building the cache daily independently will prevent this issue. The only
drawback is for the rare cases when we install new software or scripts,
the doc won't be in the list until the next rebuild (but we can always
use man directly anyway)