Duplicate all XF86 key bindings in sway with Fn equivalents
Similar as was done in commit ea70f4f6cfc7878ca5903878ffe36adef1a3cc45
for the microphone toggle.

This gives access to all actions even if the laptop is missing some XF86
keys and also provides the same location for them. We should try to use
the Fn keys primarily and keep the XF86 ones as fallback.