Update khard address books location
Address books created on Radicale use an UUID instead of a name as
identifier. We could rename them manually on the server but it causes an
issue with calendars if we want to name one of them the same as an
address book (currently had the "general" conflict) because they are
stored in the same directory.