Use popup for tmux extract tokens
One one hand it is better to have it in a pane to have the context from the
window or pane still visible. But on the other hand, the pane is taking
some space and some content is hidden anyway during the shifting.

Having it in a popup is better for the cases where we try to extract
from a small pane and there isn't much space for the selection.