Remove contrast calculations in test-contrast utility
There are a lot of controversy over the usefulness of the WCAG 2.0
especially regarding the calculations when using a dark mode.
This was already visible during the building of the theme where some
contrast values were not marked as good despite looking fine and vice
versa.
We'll just rely on the eye then for now.