Use Qt pinentry in gnupg agent
GTK2 is outdated (and looking ugly), it's only still available on the
system because of GIMP.

Once it is migrated to GTK3 however, the pinentry won't work anymore. So
we're already moving it to the Qt one which works because we have the
qt5-base dependency from qutebrowser.

We could use the GTK3 one (pinentry-gnome3) but it requires adding a
dependency in the system to gcr. Maybe later.