Sync locally created maildir folder to remote via mbsync
Removal doesn't seem to work: if I delete a folder, it is recreated
again on the next sync. mbsync can't detect it was deleted vs it is
missing, maybe because of the use of "SyncState *" that keeps track of
the syncing within the folders themselves.
Not too inconvenient for now, we'll delete them remotely.